Do not load files with relative paths
Currently, resource files (i.e. rule definition files), are loaded with
relative paths. This caused the script to look for resources in the
directory where it was executed, leading to FileNotFoundError if the
script was run outside the `src` directory. Hence, introduce `PathUtils`
and use `getResourcePath` to retrieve absolute path for the resource.
In this way, the binary script can now be executed from anywhere.

Signed-off-by: hoang1007 <hoangvuhuy1007@gmail.com>